+++ services/identity-service/src/main/java/io/ciphernance/identity/application/command/user/mfa/confirm/ConfirmMfaCommand.java
@@ -1,4 +1,4 @@
-package io.ciphernance.identity.application.command.user.mfa;
+package io.ciphernance.identity.application.command.user.mfa.confirm;
 
 import io.ciphernance.identity.application.mediator.Command;
