+++ services/identity-service/src/main/java/io/ciphernance/identity/application/command/user/mfa/confirm/ConfirmMfaHandler.java
@@ -1,58 +1,66 @@
-package io.ciphernance.identity.application.command.user.mfa;
+package io.ciphernance.identity.application.command.user.mfa.confirm;
 
 import io.ciphernance.identity.application.exception.user.InvalidMfaCodeException;
+import io.ciphernance.identity.application.exception.user.MfaSetupExpiredException;
 import io.ciphernance.identity.application.exception.user.UserNotFoundException;
 import io.ciphernance.identity.application.mediator.CommandHandler;
 import io.ciphernance.identity.application.port.out.EventPublisherPort;
+import io.ciphernance.identity.application.port.out.MfaSetupCachePort;
 import io.ciphernance.identity.application.port.out.TotpValidatorPort;
 import io.ciphernance.identity.domain.event.DomainEvent;
-import io.ciphernance.identity.domain.event.MfaDisabledEvent;
-import io.ciphernance.identity.domain.exception.MfaNotEnabledException;
+import io.ciphernance.identity.domain.event.MfaEnabledEvent;
 import io.ciphernance.identity.domain.model.User;
 import io.ciphernance.identity.domain.port.UserRepositoryPort;
 import org.springframework.stereotype.Component;
 
 @Component
-public class DisableMfaHandler implements CommandHandler<DisableMfaCommand, Void> {
+public class ConfirmMfaHandler implements CommandHandler<ConfirmMfaCommand, ConfirmMfaResponse> {
 
     private final UserRepositoryPort userRepository;
     private final TotpValidatorPort totpValidator;
+    private final MfaSetupCachePort mfaSetupCache;
     private final EventPublisherPort eventPublisher;
 
-    public DisableMfaHandler(UserRepositoryPort userRepository,
+    public ConfirmMfaHandler(UserRepositoryPort userRepository,
                              TotpValidatorPort totpValidator,
+                             MfaSetupCachePort mfaSetupCache,
                              EventPublisherPort eventPublisher) {
 
         this.userRepository = userRepository;
         this.totpValidator = totpValidator;
+        this.mfaSetupCache = mfaSetupCache;
         this.eventPublisher = eventPublisher;
     }
 
     @Override
-    public Void handle(DisableMfaCommand command) {
+    public ConfirmMfaResponse handle(ConfirmMfaCommand command) {
 
         User user = userRepository.findById(command.userId())
                 .orElseThrow(() -> new UserNotFoundException(command.userId()));
 
-        if (!user.isMfaEnabled()) {
-            throw new MfaNotEnabledException(command.userId());
-        }
+        String secret = mfaSetupCache.retrieve(command.userId())
+                .orElseThrow(() -> new MfaSetupExpiredException(command.userId()));
 
-        if (!totpValidator.validate(user.getMfaSecret(), command.totpCode())) {
+        if (!totpValidator.validate(secret, command.totpCode())) {
             throw new InvalidMfaCodeException();
         }
 
-        user.disableMfa();
+        user.enableMfa(secret);
         userRepository.save(user);
 
+        mfaSetupCache.delete(command.userId());
+
         eventPublisher.publish(
-                new MfaDisabledEvent(
+                new MfaEnabledEvent(
                         DomainEvent.newEventId(),
                         user.getId(),
                         user.getUpdatedAt()
                 )
         );
 
-        return null;
+        return new ConfirmMfaResponse(
+                user.getId(),
+                user.getUpdatedAt()
+        );
     }
 }
