+++ services/identity-service/src/main/java/io/ciphernance/identity/application/command/user/mfa/disable/DisableMfaHandler.java
@@ -1,66 +1,58 @@
-package io.ciphernance.identity.application.command.user.mfa;
+package io.ciphernance.identity.application.command.user.mfa.disable;
 
 import io.ciphernance.identity.application.exception.user.InvalidMfaCodeException;
-import io.ciphernance.identity.application.exception.user.MfaSetupExpiredException;
 import io.ciphernance.identity.application.exception.user.UserNotFoundException;
 import io.ciphernance.identity.application.mediator.CommandHandler;
 import io.ciphernance.identity.application.port.out.EventPublisherPort;
-import io.ciphernance.identity.application.port.out.MfaSetupCachePort;
 import io.ciphernance.identity.application.port.out.TotpValidatorPort;
 import io.ciphernance.identity.domain.event.DomainEvent;
-import io.ciphernance.identity.domain.event.MfaEnabledEvent;
+import io.ciphernance.identity.domain.event.MfaDisabledEvent;
+import io.ciphernance.identity.domain.exception.MfaNotEnabledException;
 import io.ciphernance.identity.domain.model.User;
 import io.ciphernance.identity.domain.port.UserRepositoryPort;
 import org.springframework.stereotype.Component;
 
 @Component
-public class ConfirmMfaHandler implements CommandHandler<ConfirmMfaCommand, ConfirmMfaResponse> {
+public class DisableMfaHandler implements CommandHandler<DisableMfaCommand, Void> {
 
     private final UserRepositoryPort userRepository;
     private final TotpValidatorPort totpValidator;
-    private final MfaSetupCachePort mfaSetupCache;
     private final EventPublisherPort eventPublisher;
 
-    public ConfirmMfaHandler(UserRepositoryPort userRepository,
+    public DisableMfaHandler(UserRepositoryPort userRepository,
                              TotpValidatorPort totpValidator,
-                             MfaSetupCachePort mfaSetupCache,
                              EventPublisherPort eventPublisher) {
 
         this.userRepository = userRepository;
         this.totpValidator = totpValidator;
-        this.mfaSetupCache = mfaSetupCache;
         this.eventPublisher = eventPublisher;
     }
 
     @Override
-    public ConfirmMfaResponse handle(ConfirmMfaCommand command) {
+    public Void handle(DisableMfaCommand command) {
 
         User user = userRepository.findById(command.userId())
                 .orElseThrow(() -> new UserNotFoundException(command.userId()));
 
-        String secret = mfaSetupCache.retrieve(command.userId())
-                .orElseThrow(() -> new MfaSetupExpiredException(command.userId()));
+        if (!user.isMfaEnabled()) {
+            throw new MfaNotEnabledException(command.userId());
+        }
 
-        if (!totpValidator.validate(secret, command.totpCode())) {
+        if (!totpValidator.validate(user.getMfaSecret(), command.totpCode())) {
             throw new InvalidMfaCodeException();
         }
 
-        user.enableMfa(secret);
+        user.disableMfa();
         userRepository.save(user);
 
-        mfaSetupCache.delete(command.userId());
-
         eventPublisher.publish(
-                new MfaEnabledEvent(
+                new MfaDisabledEvent(
                         DomainEvent.newEventId(),
                         user.getId(),
                         user.getUpdatedAt()
                 )
         );
 
-        return new ConfirmMfaResponse(
-                user.getId(),
-                user.getUpdatedAt()
-        );
+        return null;
     }
 }
